pyproject.toml
[tool.poetry]
name = "asterisk"
version = "0.1.0"
description = ""
authors = ["Jerzy Kapciak <jerzyryszardkapciak@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
sentry-sdk = "^2.7.1"
djangorestframework = "^3.15.2"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
whitenoise = "^6.7.0"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
django-phonenumber-field = "^8.0.0"
django-phonenumbers = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
